zebra: convert to mgmtd
