+++ mgmtd/mgmt_vty.c
@@ -26,6 +26,7 @@
 #include "ripd/rip_nb.h"
 #include "ripngd/ripng_nb.h"
 #include "staticd/static_vty.h"
+#include "zebra/zebra_cli.h"
 
 extern struct frr_daemon_info *mgmt_daemon_info;
 
@@ -579,6 +580,7 @@ void mgmt_vty_init(void)
 	 * backend components that are moved to new MGMTD infra
 	 * here one by one.
 	 */
+	zebra_cli_init();
 #if HAVE_RIPD
 	rip_cli_init();
 #endif
