+++ mgmtd/subdir.am
@@ -16,6 +16,7 @@ clippy_scan += \
 
 lib_LTLIBRARIES += mgmtd/libmgmt_be_nb.la
 mgmtd_libmgmt_be_nb_la_SOURCES = \
+	zebra/zebra_cli.c \
 	# end
 nodist_mgmtd_libmgmt_be_nb_la_SOURCES = \
 	lib/affinitymap_cli.c \
@@ -44,6 +45,7 @@ noinst_HEADERS += \
 	mgmtd/mgmt_history.h \
 	mgmtd/mgmt_memory.h \
 	mgmtd/mgmt_txn.h \
+	zebra/zebra_cli.h \
 	# end
 
 sbin_PROGRAMS += mgmtd/mgmtd
@@ -52,6 +54,7 @@ mgmtd_mgmtd_SOURCES = \
 	mgmtd/mgmt_main.c \
 	# end
 nodist_mgmtd_mgmtd_SOURCES = \
+	yang/frr-zebra.yang.c \
 	# nothing
 mgmtd_mgmtd_CFLAGS = $(AM_CFLAGS) -I ./
 mgmtd_mgmtd_LDADD = mgmtd/libmgmtd.a lib/libfrr.la $(LIBCAP) $(LIBM) $(LIBYANG_LIBS) $(UST_LIBS)
