+++ tests/topotests/mgmt_oper/simple-results/result-intf-eth0-only-config.json
@@ -3,7 +3,15 @@
 		"interface": [
 			{
 				"name": "r1-eth0",
-				"description": "r1-eth0-desc"
+				"description": "r1-eth0-desc",
+				"frr-zebra:zebra": {
+					"ipv4-addrs": [
+						{
+							"ip": "1.1.1.1",
+							"prefix-length": 24
+						}
+					]
+				}
 			}
 		]
 	}
