+++ tests/topotests/mgmt_oper/simple-results/result-intf-eth0-with-config.json
@@ -14,6 +14,12 @@
 					"phy-address": "rubout"
 				},
 				"frr-zebra:zebra": {
+					"ipv4-addrs": [
+						{
+							"ip": "1.1.1.1",
+							"prefix-length": 24
+						}
+					],
 					"state": {
 						"up-count": 0,
 						"down-count": 0
