+++ zebra/rtadv.c
@@ -21,7 +21,6 @@
 #include "vrf.h"
 #include "ns.h"
 #include "lib_errors.h"
-#include "northbound_cli.h"
 
 #include "zebra/interface.h"
 #include "zebra/rtadv.h"
@@ -31,7 +30,6 @@
 #include "zebra/zebra_vrf.h"
 #include "zebra/zebra_errors.h"
 #include "zebra/zebra_router.h"
-#include "zebra/zebra_nb.h"
 
 extern struct zebra_privs_t zserv_privs;
 
@@ -1550,562 +1548,6 @@ DEFPY(show_ipv6_nd_ra_if, show_ipv6_nd_ra_if_cmd,
 	return CMD_SUCCESS;
 }
 
-DEFPY_YANG (ipv6_nd_ra_fast_retrans,
-	ipv6_nd_ra_fast_retrans_cmd,
-	"[no] ipv6 nd ra-fast-retrans",
-	NO_STR
-	"Interface IPv6 config commands\n"
-	"Neighbor discovery\n"
-	"Fast retransmit of RA packets\n")
-{
-	if (no)
-		nb_cli_enqueue_change(vty,
-				      "./frr-zebra:zebra/ipv6-router-advertisements/fast-retransmit",
-				      NB_OP_MODIFY, "false");
-	else
-		nb_cli_enqueue_change(vty,
-				      "./frr-zebra:zebra/ipv6-router-advertisements/fast-retransmit",
-				      NB_OP_DESTROY, NULL);
-	return nb_cli_apply_changes(vty, NULL);
-}
-
-void lib_interface_zebra_ipv6_router_advertisements_fast_retransmit_cli_write(
-	struct vty *vty, const struct lyd_node *dnode, bool show_defaults)
-{
-	bool fast_retransmit = yang_dnode_get_bool(dnode, NULL);
-
-	if (!fast_retransmit)
-		vty_out(vty, " no ipv6 nd ra-fast-retrans\n");
-	else if (show_defaults)
-		vty_out(vty, " ipv6 nd ra-fast-retrans\n");
-}
-
-DEFPY_YANG (ipv6_nd_ra_hop_limit,
-       ipv6_nd_ra_hop_limit_cmd,
-       "[no] ipv6 nd ra-hop-limit ![(0-255)$hopcount]",
-       NO_STR
-       "Interface IPv6 config commands\n"
-       "Neighbor discovery\n"
-       "Advertisement Hop Limit\n"
-       "Advertisement Hop Limit in hops (default:64)\n")
-{
-	if (!no)
-		nb_cli_enqueue_change(vty,
-				      "./frr-zebra:zebra/ipv6-router-advertisements/cur-hop-limit",
-				      NB_OP_MODIFY, hopcount_str);
-	else
-		nb_cli_enqueue_change(vty,
-				      "./frr-zebra:zebra/ipv6-router-advertisements/cur-hop-limit",
-				      NB_OP_DESTROY, NULL);
-	return nb_cli_apply_changes(vty, NULL);
-}
-
-void lib_interface_zebra_ipv6_router_advertisements_cur_hop_limit_cli_write(
-	struct vty *vty, const struct lyd_node *dnode, bool show_defaults)
-{
-	uint8_t hop_limit = yang_dnode_get_uint8(dnode, NULL);
-
-	vty_out(vty, " ipv6 nd ra-hop-limit %u\n", hop_limit);
-}
-
-DEFPY_YANG (ipv6_nd_ra_retrans_interval,
-       ipv6_nd_ra_retrans_interval_cmd,
-       "[no] ipv6 nd ra-retrans-interval ![(0-4294967295)$interval]",
-       NO_STR
-       "Interface IPv6 config commands\n"
-       "Neighbor discovery\n"
-       "Advertisement Retransmit Interval\n"
-       "Advertisement Retransmit Interval in msec\n")
-{
-	if (!no)
-		nb_cli_enqueue_change(vty,
-				      "./frr-zebra:zebra/ipv6-router-advertisements/retrans-timer",
-				      NB_OP_MODIFY, interval_str);
-	else
-		nb_cli_enqueue_change(vty,
-				      "./frr-zebra:zebra/ipv6-router-advertisements/retrans-timer",
-				      NB_OP_DESTROY, NULL);
-	return nb_cli_apply_changes(vty, NULL);
-}
-
-void lib_interface_zebra_ipv6_router_advertisements_retrans_timer_cli_write(
-	struct vty *vty, const struct lyd_node *dnode, bool show_defaults)
-{
-	uint32_t retrans_timer = yang_dnode_get_uint32(dnode, NULL);
-
-	vty_out(vty, " ipv6 nd ra-retrans-interval %u\n", retrans_timer);
-}
-
-DEFPY_YANG (ipv6_nd_suppress_ra,
-       ipv6_nd_suppress_ra_cmd,
-       "[no] ipv6 nd suppress-ra",
-       NO_STR
-       "Interface IPv6 config commands\n"
-       "Neighbor discovery\n"
-       "Suppress Router Advertisement\n")
-{
-	if (no)
-		nb_cli_enqueue_change(vty,
-				      "./frr-zebra:zebra/ipv6-router-advertisements/send-advertisements",
-				      NB_OP_MODIFY, "true");
-	else
-		nb_cli_enqueue_change(vty,
-				      "./frr-zebra:zebra/ipv6-router-advertisements/send-advertisements",
-				      NB_OP_DESTROY, NULL);
-	return nb_cli_apply_changes(vty, NULL);
-}
-
-void lib_interface_zebra_ipv6_router_advertisements_send_advertisements_cli_write(
-	struct vty *vty, const struct lyd_node *dnode, bool show_defaults)
-{
-	bool send_advertisements = yang_dnode_get_bool(dnode, NULL);
-
-	if (send_advertisements)
-		vty_out(vty, " no ipv6 nd suppress-ra\n");
-	else if (show_defaults)
-		vty_out(vty, " ipv6 nd suppress-ra\n");
-}
-
-DEFPY_YANG (ipv6_nd_ra_interval,
-       ipv6_nd_ra_interval_cmd,
-       "[no] ipv6 nd ra-interval ![<(1-1800)$sec|msec (70-1800000)$msec>]",
-       NO_STR
-       "Interface IPv6 config commands\n"
-       "Neighbor discovery\n"
-       "Router Advertisement interval\n"
-       "Router Advertisement interval in seconds\n"
-       "Router Advertisement interval in milliseconds\n"
-       "Router Advertisement interval in milliseconds\n")
-{
-	char value[YANG_VALUE_MAXLEN];
-
-	if (!no) {
-		if (sec)
-			snprintf(value, sizeof(value), "%lu", sec * 1000);
-		else
-			snprintf(value, sizeof(value), "%lu", msec);
-
-		nb_cli_enqueue_change(vty,
-				      "./frr-zebra:zebra/ipv6-router-advertisements/max-rtr-adv-interval",
-				      NB_OP_MODIFY, value);
-	} else {
-		nb_cli_enqueue_change(vty,
-				      "./frr-zebra:zebra/ipv6-router-advertisements/max-rtr-adv-interval",
-				      NB_OP_DESTROY, NULL);
-	}
-	return nb_cli_apply_changes(vty, NULL);
-}
-
-void lib_interface_zebra_ipv6_router_advertisements_max_rtr_adv_interval_cli_write(
-	struct vty *vty, const struct lyd_node *dnode, bool show_defaults)
-{
-	uint32_t max_rtr_adv_interval = yang_dnode_get_uint32(dnode, NULL);
-
-	if (max_rtr_adv_interval % 1000)
-		vty_out(vty, " ipv6 nd ra-interval msec %u\n",
-			max_rtr_adv_interval);
-	else
-		vty_out(vty, " ipv6 nd ra-interval %u\n",
-			max_rtr_adv_interval / 1000);
-}
-
-DEFPY_YANG (ipv6_nd_ra_lifetime,
-       ipv6_nd_ra_lifetime_cmd,
-       "[no] ipv6 nd ra-lifetime ![(0-9000)$lifetime]",
-       NO_STR
-       "Interface IPv6 config commands\n"
-       "Neighbor discovery\n"
-       "Router lifetime\n"
-       "Router lifetime in seconds (0 stands for a non-default gw)\n")
-{
-	if (!no)
-		nb_cli_enqueue_change(vty,
-				      "./frr-zebra:zebra/ipv6-router-advertisements/default-lifetime",
-				      NB_OP_MODIFY, lifetime_str);
-	else
-		nb_cli_enqueue_change(vty,
-				      "./frr-zebra:zebra/ipv6-router-advertisements/default-lifetime",
-				      NB_OP_DESTROY, NULL);
-	return nb_cli_apply_changes(vty, NULL);
-}
-
-void lib_interface_zebra_ipv6_router_advertisements_default_lifetime_cli_write(
-	struct vty *vty, const struct lyd_node *dnode, bool show_defaults)
-{
-	uint16_t default_lifetime = yang_dnode_get_uint16(dnode, NULL);
-
-	vty_out(vty, " ipv6 nd ra-lifetime %u\n", default_lifetime);
-}
-
-DEFPY_YANG (ipv6_nd_reachable_time,
-       ipv6_nd_reachable_time_cmd,
-       "[no] ipv6 nd reachable-time ![(1-3600000)$msec]",
-       NO_STR
-       "Interface IPv6 config commands\n"
-       "Neighbor discovery\n"
-       "Reachable time\n"
-       "Reachable time in milliseconds\n")
-{
-	if (!no)
-		nb_cli_enqueue_change(vty,
-				      "./frr-zebra:zebra/ipv6-router-advertisements/reachable-time",
-				      NB_OP_MODIFY, msec_str);
-	else
-		nb_cli_enqueue_change(vty,
-				      "./frr-zebra:zebra/ipv6-router-advertisements/reachable-time",
-				      NB_OP_DESTROY, NULL);
-	return nb_cli_apply_changes(vty, NULL);
-}
-
-void lib_interface_zebra_ipv6_router_advertisements_reachable_time_cli_write(
-	struct vty *vty, const struct lyd_node *dnode, bool show_defaults)
-{
-	uint32_t reachable_time = yang_dnode_get_uint32(dnode, NULL);
-
-	vty_out(vty, " ipv6 nd reachable-time %u\n", reachable_time);
-}
-
-DEFPY_YANG (ipv6_nd_homeagent_preference,
-       ipv6_nd_homeagent_preference_cmd,
-       "[no] ipv6 nd home-agent-preference ![(0-65535)$pref]",
-       NO_STR
-       "Interface IPv6 config commands\n"
-       "Neighbor discovery\n"
-       "Home Agent preference\n"
-       "preference value (default is 0, least preferred)\n")
-{
-	if (!no)
-		nb_cli_enqueue_change(vty,
-				      "./frr-zebra:zebra/ipv6-router-advertisements/home-agent-preference",
-				      NB_OP_MODIFY, pref_str);
-	else
-		nb_cli_enqueue_change(vty,
-				      "./frr-zebra:zebra/ipv6-router-advertisements/home-agent-preference",
-				      NB_OP_DESTROY, NULL);
-	return nb_cli_apply_changes(vty, NULL);
-}
-
-void lib_interface_zebra_ipv6_router_advertisements_home_agent_preference_cli_write(
-	struct vty *vty, const struct lyd_node *dnode, bool show_defaults)
-{
-	uint16_t home_agent_preference = yang_dnode_get_uint16(dnode, NULL);
-
-	vty_out(vty, " ipv6 nd home-agent-preference %u\n",
-		home_agent_preference);
-}
-
-DEFPY_YANG (ipv6_nd_homeagent_lifetime,
-       ipv6_nd_homeagent_lifetime_cmd,
-       "[no] ipv6 nd home-agent-lifetime ![(1-65520)$lifetime]",
-       NO_STR
-       "Interface IPv6 config commands\n"
-       "Neighbor discovery\n"
-       "Home Agent lifetime\n"
-       "Home Agent lifetime in seconds\n")
-{
-	if (!no)
-		nb_cli_enqueue_change(vty,
-				      "./frr-zebra:zebra/ipv6-router-advertisements/home-agent-lifetime",
-				      NB_OP_MODIFY, lifetime_str);
-	else
-		nb_cli_enqueue_change(vty,
-				      "./frr-zebra:zebra/ipv6-router-advertisements/home-agent-lifetime",
-				      NB_OP_DESTROY, NULL);
-	return nb_cli_apply_changes(vty, NULL);
-}
-
-void lib_interface_zebra_ipv6_router_advertisements_home_agent_lifetime_cli_write(
-	struct vty *vty, const struct lyd_node *dnode, bool show_defaults)
-{
-	uint16_t home_agent_lifetime = yang_dnode_get_uint16(dnode, NULL);
-
-	vty_out(vty, " ipv6 nd home-agent-lifetime %u\n", home_agent_lifetime);
-}
-
-DEFPY_YANG (ipv6_nd_managed_config_flag,
-       ipv6_nd_managed_config_flag_cmd,
-       "[no] ipv6 nd managed-config-flag",
-       NO_STR
-       "Interface IPv6 config commands\n"
-       "Neighbor discovery\n"
-       "Managed address configuration flag\n")
-{
-	if (!no)
-		nb_cli_enqueue_change(vty,
-				      "./frr-zebra:zebra/ipv6-router-advertisements/managed-flag",
-				      NB_OP_MODIFY, "true");
-	else
-		nb_cli_enqueue_change(vty,
-				      "./frr-zebra:zebra/ipv6-router-advertisements/managed-flag",
-				      NB_OP_DESTROY, NULL);
-	return nb_cli_apply_changes(vty, NULL);
-}
-
-void lib_interface_zebra_ipv6_router_advertisements_managed_flag_cli_write(
-	struct vty *vty, const struct lyd_node *dnode, bool show_defaults)
-{
-	bool managed_flag = yang_dnode_get_bool(dnode, NULL);
-
-	if (managed_flag)
-		vty_out(vty, " ipv6 nd managed-config-flag\n");
-	else if (show_defaults)
-		vty_out(vty, " no ipv6 nd managed-config-flag\n");
-}
-
-DEFPY_YANG (ipv6_nd_homeagent_config_flag,
-       ipv6_nd_homeagent_config_flag_cmd,
-       "[no] ipv6 nd home-agent-config-flag",
-       NO_STR
-       "Interface IPv6 config commands\n"
-       "Neighbor discovery\n"
-       "Home Agent configuration flag\n")
-{
-	if (!no)
-		nb_cli_enqueue_change(vty,
-				      "./frr-zebra:zebra/ipv6-router-advertisements/home-agent-flag",
-				      NB_OP_MODIFY, "true");
-	else
-		nb_cli_enqueue_change(vty,
-				      "./frr-zebra:zebra/ipv6-router-advertisements/home-agent-flag",
-				      NB_OP_DESTROY, NULL);
-	return nb_cli_apply_changes(vty, NULL);
-}
-
-void lib_interface_zebra_ipv6_router_advertisements_home_agent_flag_cli_write(
-	struct vty *vty, const struct lyd_node *dnode, bool show_defaults)
-{
-	bool home_agent_flag = yang_dnode_get_bool(dnode, NULL);
-
-	if (home_agent_flag)
-		vty_out(vty, " ipv6 nd home-agent-config-flag\n");
-	else if (show_defaults)
-		vty_out(vty, " no ipv6 nd home-agent-config-flag\n");
-}
-
-DEFPY_YANG (ipv6_nd_adv_interval_config_option,
-       ipv6_nd_adv_interval_config_option_cmd,
-       "[no] ipv6 nd adv-interval-option",
-       NO_STR
-       "Interface IPv6 config commands\n"
-       "Neighbor discovery\n"
-       "Advertisement Interval Option\n")
-{
-	if (!no)
-		nb_cli_enqueue_change(vty,
-				      "./frr-zebra:zebra/ipv6-router-advertisements/advertisement-interval-option",
-				      NB_OP_MODIFY, "true");
-	else
-		nb_cli_enqueue_change(vty,
-				      "./frr-zebra:zebra/ipv6-router-advertisements/advertisement-interval-option",
-				      NB_OP_DESTROY, NULL);
-	return nb_cli_apply_changes(vty, NULL);
-}
-
-void lib_interface_zebra_ipv6_router_advertisements_advertisement_interval_option_cli_write(
-	struct vty *vty, const struct lyd_node *dnode, bool show_defaults)
-{
-	bool advertisement_interval_option = yang_dnode_get_bool(dnode, NULL);
-
-	if (advertisement_interval_option)
-		vty_out(vty, " ipv6 nd adv-interval-option\n");
-	else if (show_defaults)
-		vty_out(vty, " no ipv6 nd adv-interval-option\n");
-}
-
-DEFPY_YANG (ipv6_nd_other_config_flag,
-       ipv6_nd_other_config_flag_cmd,
-       "[no] ipv6 nd other-config-flag",
-       NO_STR
-       "Interface IPv6 config commands\n"
-       "Neighbor discovery\n"
-       "Other statefull configuration flag\n")
-{
-	if (!no)
-		nb_cli_enqueue_change(vty,
-				      "./frr-zebra:zebra/ipv6-router-advertisements/other-config-flag",
-				      NB_OP_MODIFY, "true");
-	else
-		nb_cli_enqueue_change(vty,
-				      "./frr-zebra:zebra/ipv6-router-advertisements/other-config-flag",
-				      NB_OP_DESTROY, NULL);
-	return nb_cli_apply_changes(vty, NULL);
-}
-
-void lib_interface_zebra_ipv6_router_advertisements_other_config_flag_cli_write(
-	struct vty *vty, const struct lyd_node *dnode, bool show_defaults)
-{
-	bool other_config_flag = yang_dnode_get_bool(dnode, NULL);
-
-	if (other_config_flag)
-		vty_out(vty, " ipv6 nd other-config-flag\n");
-	else if (show_defaults)
-		vty_out(vty, " no ipv6 nd other-config-flag\n");
-}
-
-DEFPY_YANG (ipv6_nd_prefix,
-       ipv6_nd_prefix_cmd,
-       "[no] ipv6 nd prefix X:X::X:X/M$prefix [<(0-4294967295)|infinite>$valid <(0-4294967295)|infinite>$preferred] [{router-address$routeraddr|off-link$offlink|no-autoconfig$noautoconf}]",
-       NO_STR
-       "Interface IPv6 config commands\n"
-       "Neighbor discovery\n"
-       "Prefix information\n"
-       "IPv6 prefix\n"
-       "Valid lifetime in seconds\n"
-       "Infinite valid lifetime\n"
-       "Preferred lifetime in seconds\n"
-       "Infinite preferred lifetime\n"
-       "Set Router Address flag\n"
-       "Do not use prefix for onlink determination\n"
-       "Do not use prefix for autoconfiguration\n")
-{
-	if (!no) {
-		nb_cli_enqueue_change(vty, ".", NB_OP_CREATE, NULL);
-		if (valid) {
-			if (strmatch(valid, "infinite"))
-				valid = "4294967295";
-			nb_cli_enqueue_change(vty, "./valid-lifetime",
-					      NB_OP_MODIFY, valid);
-		} else {
-			nb_cli_enqueue_change(vty, "./valid-lifetime",
-					      NB_OP_DESTROY, NULL);
-		}
-		if (preferred) {
-			if (strmatch(preferred, "infinite"))
-				preferred = "4294967295";
-			nb_cli_enqueue_change(vty, "./preferred-lifetime",
-					      NB_OP_MODIFY, preferred);
-		} else {
-			nb_cli_enqueue_change(vty, "./preferred-lifetime",
-					      NB_OP_DESTROY, NULL);
-		}
-		if (routeraddr)
-			nb_cli_enqueue_change(vty, "./router-address-flag",
-					      NB_OP_MODIFY, "true");
-		else
-			nb_cli_enqueue_change(vty, "./router-address-flag",
-					      NB_OP_DESTROY, NULL);
-		if (offlink)
-			nb_cli_enqueue_change(vty, "./on-link-flag",
-					      NB_OP_MODIFY, "false");
-		else
-			nb_cli_enqueue_change(vty, "./on-link-flag",
-					      NB_OP_DESTROY, NULL);
-		if (noautoconf)
-			nb_cli_enqueue_change(vty, "./autonomous-flag",
-					      NB_OP_MODIFY, "false");
-		else
-			nb_cli_enqueue_change(vty, "./autonomous-flag",
-					      NB_OP_DESTROY, NULL);
-	} else {
-		nb_cli_enqueue_change(vty, ".", NB_OP_DESTROY, NULL);
-	}
-	return nb_cli_apply_changes(
-		vty,
-		"./frr-zebra:zebra/ipv6-router-advertisements/prefix-list/prefix[prefix-spec='%s']",
-		prefix_str);
-}
-
-void lib_interface_zebra_ipv6_router_advertisements_prefix_list_prefix_cli_write(
-	struct vty *vty, const struct lyd_node *dnode, bool show_defaults)
-{
-	const char *prefix = yang_dnode_get_string(dnode, "prefix-spec");
-	struct lyd_node *valid = yang_dnode_get(dnode, "valid-lifetime");
-	struct lyd_node *preferred = yang_dnode_get(dnode, "preferred-lifetime");
-	bool router_address_flag = yang_dnode_get_bool(dnode,
-						       "router-address-flag");
-	bool on_link_flag = yang_dnode_get_bool(dnode, "on-link-flag");
-	bool autonomous_flag = yang_dnode_get_bool(dnode, "autonomous-flag");
-
-	vty_out(vty, " ipv6 nd prefix %s", prefix);
-
-	if (!yang_dnode_is_default(valid, NULL) ||
-	    !yang_dnode_is_default(preferred, NULL) || show_defaults) {
-		uint32_t valid_lifetime = yang_dnode_get_uint32(valid, NULL);
-		uint32_t preferred_lifetime = yang_dnode_get_uint32(preferred,
-								    NULL);
-
-		if (valid_lifetime == UINT32_MAX)
-			vty_out(vty, " infinite");
-		else
-			vty_out(vty, " %u", valid_lifetime);
-		if (preferred_lifetime == UINT32_MAX)
-			vty_out(vty, " infinite");
-		else
-			vty_out(vty, " %u", preferred_lifetime);
-	}
-
-	if (!on_link_flag)
-		vty_out(vty, " off-link");
-
-	if (!autonomous_flag)
-		vty_out(vty, " no-autoconfig");
-
-	if (router_address_flag)
-		vty_out(vty, " router-address");
-
-	vty_out(vty, "\n");
-}
-
-DEFPY_YANG (ipv6_nd_router_preference,
-       ipv6_nd_router_preference_cmd,
-       "[no] ipv6 nd router-preference ![<high|medium|low>$pref]",
-       NO_STR
-       "Interface IPv6 config commands\n"
-       "Neighbor discovery\n"
-       "Default router preference\n"
-       "High default router preference\n"
-       "Medium default router preference (default)\n"
-       "Low default router preference\n")
-{
-	if (!no)
-		nb_cli_enqueue_change(vty,
-				      "./frr-zebra:zebra/ipv6-router-advertisements/default-router-preference",
-				      NB_OP_MODIFY, pref);
-	else
-		nb_cli_enqueue_change(vty,
-				      "./frr-zebra:zebra/ipv6-router-advertisements/default-router-preference",
-				      NB_OP_DESTROY, NULL);
-	return nb_cli_apply_changes(vty, NULL);
-}
-
-void lib_interface_zebra_ipv6_router_advertisements_default_router_preference_cli_write(
-	struct vty *vty, const struct lyd_node *dnode, bool show_defaults)
-{
-	const char *default_router_preference = yang_dnode_get_string(dnode,
-								      NULL);
-
-	vty_out(vty, " ipv6 nd router-preference %s\n",
-		default_router_preference);
-}
-
-DEFPY_YANG (ipv6_nd_mtu,
-       ipv6_nd_mtu_cmd,
-       "[no] ipv6 nd mtu ![(1-65535)]",
-       NO_STR
-       "Interface IPv6 config commands\n"
-       "Neighbor discovery\n"
-       "Advertised MTU\n"
-       "MTU in bytes\n")
-{
-	if (!no)
-		nb_cli_enqueue_change(vty,
-				      "./frr-zebra:zebra/ipv6-router-advertisements/link-mtu",
-				      NB_OP_MODIFY, mtu_str);
-	else
-		nb_cli_enqueue_change(vty,
-				      "./frr-zebra:zebra/ipv6-router-advertisements/link-mtu",
-				      NB_OP_DESTROY, NULL);
-	return nb_cli_apply_changes(vty, NULL);
-}
-
-void lib_interface_zebra_ipv6_router_advertisements_link_mtu_cli_write(
-	struct vty *vty, const struct lyd_node *dnode, bool show_defaults)
-{
-	uint16_t link_mtu = yang_dnode_get_uint32(dnode, NULL);
-
-	vty_out(vty, " ipv6 nd mtu %u\n", link_mtu);
-}
-
 static struct rtadv_rdnss *rtadv_rdnss_new(void)
 {
 	return XCALLOC(MTYPE_RTADV_RDNSS, sizeof(struct rtadv_rdnss));
@@ -2199,123 +1641,6 @@ int rtadv_dnssl_encode(uint8_t *out, const char *in)
 
 	out[outp++] = '\0';
 	return outp;
-}
-
-DEFPY_YANG (ipv6_nd_rdnss,
-      ipv6_nd_rdnss_cmd,
-      "[no] ipv6 nd rdnss X:X::X:X$addr [<(0-4294967295)|infinite>]$lifetime",
-      NO_STR
-      "Interface IPv6 config commands\n"
-      "Neighbor discovery\n"
-      "Recursive DNS server information\n"
-      "IPv6 address\n"
-      "Valid lifetime in seconds\n"
-      "Infinite valid lifetime\n")
-{
-	if (!no) {
-		nb_cli_enqueue_change(vty, ".", NB_OP_CREATE, NULL);
-		if (lifetime) {
-			if (strmatch(lifetime, "infinite"))
-				lifetime = "4294967295";
-			nb_cli_enqueue_change(vty, "./lifetime", NB_OP_MODIFY,
-					      lifetime);
-		} else {
-			nb_cli_enqueue_change(vty, "./lifetime", NB_OP_DESTROY,
-					      NULL);
-		}
-	} else {
-		nb_cli_enqueue_change(vty, ".", NB_OP_DESTROY, NULL);
-	}
-	return nb_cli_apply_changes(
-		vty,
-		"./frr-zebra:zebra/ipv6-router-advertisements/rdnss/rdnss-address[address='%s']",
-		addr_str);
-}
-
-void lib_interface_zebra_ipv6_router_advertisements_rdnss_rdnss_address_cli_write(
-	struct vty *vty, const struct lyd_node *dnode, bool show_defaults)
-{
-	const char *address = yang_dnode_get_string(dnode, "address");
-
-	vty_out(vty, " ipv6 nd rdnss %s", address);
-
-	if (yang_dnode_exists(dnode, "lifetime")) {
-		uint32_t lifetime = yang_dnode_get_uint32(dnode, "lifetime");
-
-		if (lifetime == UINT32_MAX)
-			vty_out(vty, " infinite");
-		else
-			vty_out(vty, " %u", lifetime);
-	}
-
-	vty_out(vty, "\n");
-}
-
-DEFPY_YANG (ipv6_nd_dnssl,
-      ipv6_nd_dnssl_cmd,
-      "[no] ipv6 nd dnssl SUFFIX [<(0-4294967295)|infinite>]$lifetime",
-      NO_STR
-      "Interface IPv6 config commands\n"
-      "Neighbor discovery\n"
-      "DNS search list information\n"
-      "Domain name suffix\n"
-      "Valid lifetime in seconds\n"
-      "Infinite valid lifetime\n")
-{
-	struct rtadv_dnssl dnssl;
-	size_t len;
-
-	len = strlcpy(dnssl.name, suffix, sizeof(dnssl.name));
-	if (len == 0 || len >= sizeof(dnssl.name)) {
-		vty_out(vty, "Malformed DNS search domain\n");
-		return CMD_WARNING_CONFIG_FAILED;
-	}
-	if (dnssl.name[len - 1] == '.') {
-		/*
-		 * Allow, but don't require, a trailing dot signifying the root
-		 * zone. Canonicalize by cutting it off if present.
-		 */
-		dnssl.name[len - 1] = '\0';
-		len--;
-	}
-
-	if (!no) {
-		nb_cli_enqueue_change(vty, ".", NB_OP_CREATE, NULL);
-		if (lifetime) {
-			if (strmatch(lifetime, "infinite"))
-				lifetime = "4294967295";
-			nb_cli_enqueue_change(vty, "./lifetime", NB_OP_MODIFY,
-					      lifetime);
-		} else {
-			nb_cli_enqueue_change(vty, "./lifetime", NB_OP_DESTROY,
-					      NULL);
-		}
-	} else {
-		nb_cli_enqueue_change(vty, ".", NB_OP_DESTROY, NULL);
-	}
-	return nb_cli_apply_changes(
-		vty,
-		"./frr-zebra:zebra/ipv6-router-advertisements/dnssl/dnssl-domain[domain='%s']",
-		dnssl.name);
-}
-
-void lib_interface_zebra_ipv6_router_advertisements_dnssl_dnssl_domain_cli_write(
-	struct vty *vty, const struct lyd_node *dnode, bool show_defaults)
-{
-	const char *domain = yang_dnode_get_string(dnode, "domain");
-
-	vty_out(vty, " ipv6 nd dnssl %s", domain);
-
-	if (yang_dnode_exists(dnode, "lifetime")) {
-		uint32_t lifetime = yang_dnode_get_uint32(dnode, "lifetime");
-
-		if (lifetime == UINT32_MAX)
-			vty_out(vty, " infinite");
-		else
-			vty_out(vty, " %u", lifetime);
-	}
-
-	vty_out(vty, "\n");
 }
 
 /* Dump interface ND information to vty. */
@@ -2535,25 +1860,6 @@ void rtadv_cmd_init(void)
 	hook_register(zebra_if_extra_info, nd_dump_vty);
 
 	install_element(VIEW_NODE, &show_ipv6_nd_ra_if_cmd);
-
-	install_element(INTERFACE_NODE, &ipv6_nd_ra_fast_retrans_cmd);
-	install_element(INTERFACE_NODE, &ipv6_nd_ra_retrans_interval_cmd);
-	install_element(INTERFACE_NODE, &ipv6_nd_ra_hop_limit_cmd);
-	install_element(INTERFACE_NODE, &ipv6_nd_suppress_ra_cmd);
-	install_element(INTERFACE_NODE, &ipv6_nd_ra_interval_cmd);
-	install_element(INTERFACE_NODE, &ipv6_nd_ra_lifetime_cmd);
-	install_element(INTERFACE_NODE, &ipv6_nd_reachable_time_cmd);
-	install_element(INTERFACE_NODE, &ipv6_nd_managed_config_flag_cmd);
-	install_element(INTERFACE_NODE, &ipv6_nd_other_config_flag_cmd);
-	install_element(INTERFACE_NODE, &ipv6_nd_homeagent_config_flag_cmd);
-	install_element(INTERFACE_NODE, &ipv6_nd_homeagent_preference_cmd);
-	install_element(INTERFACE_NODE, &ipv6_nd_homeagent_lifetime_cmd);
-	install_element(INTERFACE_NODE, &ipv6_nd_adv_interval_config_option_cmd);
-	install_element(INTERFACE_NODE, &ipv6_nd_prefix_cmd);
-	install_element(INTERFACE_NODE, &ipv6_nd_router_preference_cmd);
-	install_element(INTERFACE_NODE, &ipv6_nd_mtu_cmd);
-	install_element(INTERFACE_NODE, &ipv6_nd_rdnss_cmd);
-	install_element(INTERFACE_NODE, &ipv6_nd_dnssl_cmd);
 }
 
 static int if_join_all_router(int sock, struct interface *ifp)
