+++ zebra/zebra_ptm.h
@@ -58,7 +58,6 @@ struct zebra_ptm_cb {
 void zebra_ptm_init(void);
 void zebra_ptm_finish(void);
 void zebra_ptm_connect(struct event *t);
-void zebra_ptm_write(struct vty *vty);
 int zebra_ptm_get_enable_state(void);
 
 #if HAVE_BFDD == 0
